fix: resolved the problem of creating dummy

+++ ALU.h
@@ -5,7 +5,7 @@
 
 class ALU{
     public:
-        ALU(uint32_t data1, uint32_t data2, uint32_t data3, ALUControl* aluControl, int32_t (&regFile)[32], int instruction, uint8_t (&mem)[1000000], int& pc);
+        ALU(uint32_t data1, uint32_t data2, uint32_t data3, ALUControl* aluControl, int32_t (&regFile)[32], int instruction, uint8_t (&mem)[1000000], unsigned long& pc);
         uint32_t rs1;
         uint32_t rs2;
         uint32_t rd;
@@ -14,7 +14,7 @@ class ALU{
         int immediate;
         // Might need to compute immediate value
         bool pc_write;
-        int &pc;
+        unsigned long &pc;
         int pc_target;
         ALUControl* aluControl;
         int32_t (&regs)[32];
